Rename TodoApi directory to backend
All files and directories under TodoApi have been moved to a new subdirectory called backend. This change improves the project's directory structure and makes it easier to expand the backend portion without cluttering the root directory.
