--- TodoApi/TodoApi.http
@@ -1,6 +1,0 @@
-@TodoApi_HostAddress = http://localhost:5160
-
-GET {{TodoApi_HostAddress}}/weatherforecast/
-Accept: application/json
-
-###
